Clarify `status` and `message` in test runner spec (#175)
* Clarify top-level `fail` in test runner spec

Consider the situation of:
- The user submits a solution in a compiled language
- The user's solution compiles without error
- The first test passes, but the next test "produces an error" (for
  example, it raises an exception).

In this case, the correct top-level status is `fail`, with one
test-level status of `pass`, and one test-level status of `error`. But
the previous wording could imply that a top-level status of `fail`
is only correct if at least one test has the status `fail`.

* Clarify per-test `message` in test runner spec

The previous wording could imply that a per-test `message` property
can only be present when the test `status` is `fail`.

* Clarify top-level `error` in test runner spec

To me, the previous wording implied that a top-level `status` of `error`
is valid when at least one test "errors". But actually, if at least one
test was "executed", the top-level status should be `pass` or `fail`.

* Try to clarify test-level `error` in test runner spec

* Further clarify test-level `message`

Co-authored-by: Erik Schierboom <erik_schierboom@hotmail.com>

* Describe top-level `error` more consistently

Co-authored-by: Erik Schierboom <erik_schierboom@hotmail.com>

* Update building/tooling/test-runners/interface.md

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>

Co-authored-by: Erik Schierboom <erik_schierboom@hotmail.com>