Moved and resized image
I could use some help as this is my first time trying to upload a file and reference it.

It seems images go in  docs/.imgs/. Is that right?

I couldn't get a relative web address to work. Is that right too?

I used <img src=

https://raw.githubusercontent.com/exercism/docs/79cdc57261c3c6209a16c55d3005fa7378d789e1/.imgs/architecture-min.png

Which feels like it may break.