Explain consequence of toggling config variable
The 'concept_exercises' config variable could be misunderstood to mean that
a track has concept exercises, whereas in practice it is used to toggle
learning mode on or off for an entire track.

Toggling to 'false' allows track maintainers to build out concept exercises
without making them available on the site.