Remove inconsistencies about getting help
We references to the unofficial Exercism help channels on
Gitter with references to the new official Exercism forum.

In a couple of places this introduced inconsistencies.

The first is a sample `help.md` file, where instead of
linking to the Gitter channel we referenced a language-
specific sub-channel on the forum. However this was in
a document where we discourage links to Exercism-wide
resources, which will already be included in the docs
seen by the end-user.

The second is the name of a markdown link-reference.
We were linking to the Building Exercism sub-channel,
but called the link-reference 'forum'. Elsewhere,
'forum' links to the top-level forum, and we've called
the reference 'building-exercism' when linking to
this category.

This fixes these inconsistencies.